Move testing metadata manager to test classpath
