+++ core/trino-main/src/test/java/io/trino/execution/TestRollbackTask.java
@@ -38,7 +38,7 @@ import java.util.concurrent.Future;
 import static io.airlift.concurrent.MoreFutures.getFutureValue;
 import static io.airlift.concurrent.Threads.daemonThreadsNamed;
 import static io.trino.execution.querystats.PlanOptimizersStatsCollector.createPlanOptimizersStatsCollector;
-import static io.trino.metadata.MetadataManager.createTestMetadataManager;
+import static io.trino.metadata.TestMetadataManager.createTestMetadataManager;
 import static io.trino.plugin.tpch.TpchMetadata.TINY_SCHEMA_NAME;
 import static io.trino.spi.StandardErrorCode.NOT_IN_TRANSACTION;
 import static io.trino.testing.TestingSession.testSessionBuilder;
