+++ core/trino-main/src/test/java/io/trino/execution/TestSetSessionAuthorizationTask.java
@@ -16,6 +16,7 @@ package io.trino.execution;
 import io.trino.client.NodeVersion;
 import io.trino.execution.warnings.WarningCollector;
 import io.trino.metadata.Metadata;
+import io.trino.metadata.TestMetadataManager;
 import io.trino.security.AccessControl;
 import io.trino.security.AllowAllAccessControl;
 import io.trino.spi.TrinoException;
@@ -36,7 +37,6 @@ import java.util.concurrent.ExecutorService;
 
 import static io.airlift.concurrent.Threads.daemonThreadsNamed;
 import static io.trino.execution.querystats.PlanOptimizersStatsCollector.createPlanOptimizersStatsCollector;
-import static io.trino.metadata.MetadataManager.testMetadataManagerBuilder;
 import static io.trino.testing.TestingSession.testSessionBuilder;
 import static io.trino.transaction.InMemoryTransactionManager.createTestTransactionManager;
 import static java.util.Collections.emptyList;
@@ -61,7 +61,7 @@ public class TestSetSessionAuthorizationTask
     {
         transactionManager = createTestTransactionManager();
         accessControl = new AllowAllAccessControl();
-        metadata = testMetadataManagerBuilder()
+        metadata = TestMetadataManager.builder()
                 .withTransactionManager(transactionManager)
                 .build();
         parser = new SqlParser();
