+++ core/trino-main/src/test/java/io/trino/server/TestQuerySessionSupplier.java
@@ -24,6 +24,7 @@ import io.trino.Session;
 import io.trino.connector.system.GlobalSystemConnector;
 import io.trino.metadata.Metadata;
 import io.trino.metadata.SessionPropertyManager;
+import io.trino.metadata.TestMetadataManager;
 import io.trino.security.AllowAllAccessControl;
 import io.trino.server.protocol.PreparedStatementEncoder;
 import io.trino.server.protocol.spooling.QueryDataEncoder;
@@ -45,8 +46,7 @@ import static io.trino.SystemSessionProperties.QUERY_MAX_MEMORY;
 import static io.trino.client.ProtocolHeaders.TRINO_HEADERS;
 import static io.trino.metadata.GlobalFunctionCatalog.BUILTIN_SCHEMA;
 import static io.trino.metadata.LanguageFunctionManager.QUERY_LOCAL_SCHEMA;
-import static io.trino.metadata.MetadataManager.createTestMetadataManager;
-import static io.trino.metadata.MetadataManager.testMetadataManagerBuilder;
+import static io.trino.metadata.TestMetadataManager.createTestMetadataManager;
 import static io.trino.spi.type.TimeZoneKey.getTimeZoneKey;
 import static io.trino.transaction.InMemoryTransactionManager.createTestTransactionManager;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -241,7 +241,7 @@ public class TestQuerySessionSupplier
     private static QuerySessionSupplier createSessionSupplier(SqlEnvironmentConfig config)
     {
         TransactionManager transactionManager = createTestTransactionManager();
-        Metadata metadata = testMetadataManagerBuilder()
+        Metadata metadata = TestMetadataManager.builder()
                 .withTransactionManager(transactionManager)
                 .build();
         return new QuerySessionSupplier(
